Project team management stuff

+++ Marketplace/Api/ProjectContributionController.cs
@@ -1,10 +1,13 @@
 ﻿using Marketplace.Services;
 using System;
 using System.Collections.Generic;
+using System.Dynamic;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Web;
+using System.Web.Http;
+using Umbraco.Web;
 using Umbraco.Web.WebApi;
 
 namespace Marketplace.Api
@@ -12,15 +15,72 @@ namespace Marketplace.Api
     [MemberAuthorize(AllowType = "member")]
     public class ProjectContributionController : UmbracoApiController
     {
-        public HttpResponseMessage Delete(int projectId, int memberId)
+        [HttpPut]
+        public HttpResponseMessage UpdateCollaborationStatus(int projectId, bool status)
         {
-            using (var cs = new ContributionService())
+            var cs = Services.ContentService;
+            var project = cs.GetById(projectId);
+
+            if (project.GetValue<int>("owner") == Members.GetCurrentMemberId())
             {
-                cs.DeleteContributor(projectId, memberId);
+                project.SetValue("openForCollab", status);
+
+                cs.Save(project);
+
+                return new HttpResponseMessage(HttpStatusCode.Accepted);
+            }
+            else
+            {
+                return new HttpResponseMessage(HttpStatusCode.Forbidden);
+            }
+        }
+
+        [HttpGet]
+        public ExpandoObject AddContributor(int projectId, string email)
+        {
+
+           dynamic o = new ExpandoObject();
+           var member = Members.GetByEmail(email) ;
+           if(member == null)
+           {
+               o.success = false;
+               o.error = "Email not found";
+               return o;
+
+           }
+           
+            using(var cs = new  ContributionService())
+            {
+                cs.AddContributor(projectId, member.Id);
+                o.success = true;
+                o.memberName = member.Name;
+                o.memberId = member.Id;
+                return o;
             }
 
-            var response = new HttpResponseMessage(HttpStatusCode.Accepted);
-            return response;
+
+
+
+        }
+
+        public HttpResponseMessage DeleteContributor(int projectId, int memberId)
+        {
+            UmbracoHelper help = new UmbracoHelper(UmbracoContext);
+            var project = help.TypedContent(projectId);
+
+            if(project.GetPropertyValue<int>("owner") == Members.GetCurrentMemberId())
+            { 
+                using (var cs = new ContributionService())
+                {
+                    cs.DeleteContributor(projectId, memberId);
+                }
+
+                return new HttpResponseMessage(HttpStatusCode.Accepted);
+            }
+            else
+            {
+                return new HttpResponseMessage(HttpStatusCode.Forbidden);
+            }
         }
     }
 }
